Added selector to filter watched movies instead of bookmarked ones in the Favorites tab

+++ src/app/lib/providers/favorites.js
@@ -10,7 +10,11 @@
     };
 
     var queryTorrents = function (filters) {
-        return App.db.getBookmarks(filters)
+        let query_func = App.db.getBookmarks; // default to favorites
+        if (filters.kind == 'Watched') {
+            query_func = App.db.getWatched;
+        }
+        return query_func(filters)
             .then(function (data) {
                     return data;
                 },
@@ -94,19 +98,54 @@
         return sorted;
     };
 
+    var movie_fetch_wrapper = async function (imdb_id) {
+        let movieProvider = App.Config.getProviderForType('movie')[0];
+        return movieProvider.fetch({keywords: imdb_id, page:1}).then(function (movies) {
+            if (movies.results.length !== 1) {
+                throw "Unknown movie in watched DB or cannot fetch movie data";
+            }
+            return movies.results[0];
+        });
+    };
+
+    var show_fetch_wrapper = async function (imdb_id) {
+        let showProvider = App.Config.getProviderForType('tvshow')[0];
+        return showProvider.detail(imdb_id, {
+            contextLocale: App.settings.contextLanguage || App.settings.language
+        }).then(function (show) {
+            return show;
+        });
+    };
+
     var formatForButter = function (items) {
         var movieList = [];
 
         items.forEach(function (movie) {
-
+            var movie_fetch_func = Database.getMovie;
+            var show_fetch_func = Database.getTVShowByImdb;
+            var shouldMark = true;
+            // note: in the future we could check if the movie is also in
+            // the local db to avoid fetching data we already have
+            if (isUndefined(movie.imdb_id)) {
+                shouldMark = false;
+                movie.imdb_id = movie.movie_id;
+                // if we're displaying movies from the watched database
+                // we'll fetch them from the providers instead of the local database
+                // given that storing them there would make the movie database
+                // very large
+                movie_fetch_func = movie_fetch_wrapper;
+                show_fetch_func = show_fetch_wrapper;
+            }
             var deferred = Q.defer();
             // we check if its a movie
             // or tv show then we extract right data
             if (movie.type === 'movie') {
                 // its a movie
-                Database.getMovie(movie.imdb_id)
+                movie_fetch_func(movie.imdb_id)
                     .then(function (data) {
-                            data.type = 'bookmarkedmovie';
+                            if (shouldMark) {
+                                data.type = 'bookmarkedmovie';
+                            }
                             if (/slurm.trakt.us/.test(data.image)) {
                                 data.image = data.image.replace(/slurm.trakt.us/, 'walter.trakt.us');
                             }
@@ -118,9 +157,11 @@
             } else {
                 // its a tv show
                 var _data = null;
-                Database.getTVShowByImdb(movie.imdb_id)
+                show_fetch_func(movie.imdb_id)
                     .then(function (data) {
-                        data.type = 'bookmarkedshow';
+                        if (shouldMark) {
+                            data.type = 'bookmarkedshow';
+                        }
                         data.imdb = data.imdb_id;
                         // Fallback for old bookmarks without provider in database or marked as Eztv
                         if (typeof (data.provider) === 'undefined' || data.provider === 'Eztv') {
@@ -141,12 +182,14 @@
                         deferred.reject(err);
                     }).then(function (data) {
                         if (data) {
-                            // Cache new show and return
-                            Database.deleteBookmark(_data.imdb_id);
-                            Database.deleteTVShow(_data.imdb_id);
-                            Database.addTVShow(data);
-                            Database.addBookmark(data.imdb_id, 'tvshow');
-                            data.type = 'bookmarkedshow';
+                            if (shouldMark) {
+                                // Cache new show and return
+                                Database.deleteBookmark(_data.imdb_id);
+                                Database.deleteTVShow(_data.imdb_id);
+                                Database.addTVShow(data);
+                                Database.addBookmark(data.imdb_id, 'tvshow');
+                                data.type = 'bookmarkedshow';
+                            }
                             deferred.resolve(data);
                         }
                     }, function (err) {
@@ -172,7 +215,8 @@
 
     Favorites.prototype.fetch = function (filters) {
         var params = {
-            page: filters.page
+            page: filters.page,
+            kind: filters.kind
         };
         if (filters.type === 'Series') {
             params.type = 'show';
@@ -190,10 +234,12 @@
 
     Favorites.prototype.filters = function () {
         const data = {
+            kinds: ['Favorites', 'Watched'],
             types: ['All', 'Movies', 'Series', 'Anime'],
             sorters: ['watched items', 'year', 'title', 'rating']
         };
         let filters = {
+            kinds: {},
             types: {},
             sorters: {},
         };
@@ -202,6 +248,9 @@
         }
         for (const type of data.types) {
             filters.types[type] = i18n.__(type);
+        }
+        for (const kind of data.kinds) {
+            filters.kinds[kind] = i18n.__(kind);
         }
 
         return Promise.resolve(filters);
